Coding Guidelines

+++ addons/binding/org.openhab.binding.fronius/src/main/java/org/openhab/binding/fronius/api/ValueUnit.java
@@ -1,14 +1,28 @@
+/**
+ * Copyright (c) 2010-2017 by the respective copyright holders.
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
 package org.openhab.binding.fronius.api;
 
 import com.google.gson.annotations.SerializedName;
 
+/**
+ * The {@link InverterRealtimeResponse} is responsible for storing
+ * a value
+ *
+ * @author Thomas Rokohl - Initial contribution
+ */
 public class ValueUnit {
     @SerializedName("Value")
-    private float value;
+    private double value;
     @SerializedName("Unit")
     private String unit;
 
-    public float getValue() {
+    public double getValue() {
         return value;
     }
